+++ contracts/feature-tests/basic-features/src/big_int_methods.rs
@@ -1,0 +1,5 @@
+elrond_wasm::imports!();
+
+/// TODO: add neg, abs, sqrt etc. here
+#[elrond_wasm_derive::module]
+pub trait BigIntMethods {}
