+++ contracts/feature-tests/basic-features/src/block_info_features.rs
@@ -1,0 +1,55 @@
+elrond_wasm::imports!();
+
+/// Block info getters.
+#[elrond_wasm_derive::module]
+pub trait BlockInfoFeatures {
+	#[view]
+	fn get_block_timestamp(&self) -> u64 {
+		self.blockchain().get_block_timestamp()
+	}
+
+	#[view]
+	fn get_block_nonce(&self) -> u64 {
+		self.blockchain().get_block_nonce()
+	}
+
+	#[view]
+	fn get_block_round(&self) -> u64 {
+		self.blockchain().get_block_round()
+	}
+
+	#[view]
+	fn get_block_epoch(&self) -> u64 {
+		self.blockchain().get_block_epoch()
+	}
+
+	#[view]
+	fn get_block_random_seed(&self) -> Box<[u8; 48]> {
+		self.blockchain().get_block_random_seed()
+	}
+
+	#[view]
+	fn get_prev_block_timestamp(&self) -> u64 {
+		self.blockchain().get_prev_block_timestamp()
+	}
+
+	#[view]
+	fn get_prev_block_nonce(&self) -> u64 {
+		self.blockchain().get_prev_block_nonce()
+	}
+
+	#[view]
+	fn get_prev_block_round(&self) -> u64 {
+		self.blockchain().get_prev_block_round()
+	}
+
+	#[view]
+	fn get_prev_block_epoch(&self) -> u64 {
+		self.blockchain().get_prev_block_epoch()
+	}
+
+	#[view]
+	fn get_prev_block_random_seed(&self) -> Box<[u8; 48]> {
+		self.blockchain().get_prev_block_random_seed()
+	}
+}
