+++ contracts/feature-tests/basic-features/src/types/large_boxed_byte_array.rs
