+++ contracts/feature-tests/basic-features/src/types/mod.rs
@@ -1,0 +1,9 @@
+mod large_boxed_byte_array;
+mod ser_ex1;
+mod ser_ex2;
+mod simple_enum;
+
+pub use large_boxed_byte_array::LargeBoxedByteArray;
+pub use ser_ex1::*;
+pub use ser_ex2::*;
+pub use simple_enum::*;
