+++ contracts/feature-tests/basic-features/src/types/ser_ex1.rs
