+++ contracts/feature-tests/basic-features/src/types/ser_ex2.rs
