+++ contracts/feature-tests/basic-features/src/types/simple_enum.rs
